second

+++ project/apps/payment/models.py
@@ -7,7 +7,7 @@ class Copoun(models.Model):
     product = models.ForeignKey('category.Product', models.SET_NULL, null=True, blank=True, related_name='product_copouns')
     order = models.ForeignKey('orders.Order', models.SET_NULL, null=True, blank=True, related_name='order_copouns')
     used = models.BooleanField(default=False)
-    money = models.DecimalField(max_digits=6, decimal_places=5)
+    money = models.FloatField()
     created = models.DateTimeField(auto_now_add=True)
     expire = models.DateTimeField(null=True)
     binary = models.UUIDField()
